Reorganized and extracted multiple files
