+++ src/components/pageBody/proposal/emergencyList/emergencyInfo/voting/EmergencyVoteBalance.tsx
