+++ src/components/pageBody/proposal/proposalCreationModal/proposalCreationForm/ProposalSubmittedText.tsx
