+++ src/components/pageBody/proposal/proposalCreationModal/proposalCreationForm/inputs/InputInitialFunding.tsx
