+++ src/components/pageBody/proposal/proposalCreationModal/proposalCreationForm/inputs/InputPaymentSchedule.tsx
