+++ src/components/pageBody/proposal/proposalList/proposalInfo/funding/InputFundingAmount.tsx
