+++ src/components/pageBody/proposal/proposalList/proposalInfo/voting/InputVoteAmount.tsx
