+++ src/components/pageBody/proposal/proposalList/proposalInfo/voting/VoteButtons.tsx
