+++ src/components/pageBody/proposal/proposalList/proposalInfo/voting/VoteTokenBalance.tsx
